add skin functionality

+++ Assets/Scripts/Data/PlayerData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New PlayerData", menuName = "Game/PlayerData")]
 public class PlayerData : ScriptableObject
 {
-    public int coins; // The amount of coins
+    public float coins; // The amount of coins
     public ItemData[] items; // An array of item scriptable objects
     public int inventorySize;
 }
